requirements.txt
# Core dependencies
numpy>=1.22.0
Pillow>=9.0.0
matplotlib>=3.5.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.20.0

# Object detection
ultralytics>=8.0.0  # For YOLOv8

# Image captioning
timm>=0.6.0  # Required for BLIP

# NLP for relationship extraction
spacy>=3.5.0
nltk>=3.8.0

# Graph processing and visualization
networkx>=3.0.0
pydot>=1.4.2
graphviz>=0.20.0

# Utilities
tqdm>=4.65.0
opencv-python>=4.7.0 